thêm servicer mới

+++ NT.BLL/Services/GenericService.cs
@@ -1,13 +1,14 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 using NT.BLL.Interfaces;
 
 namespace NT.BLL.Services
 {
     /// <summary>
-    /// Tri?n khai m?c ??nh cho IGenericService s? d?ng IGenericRepository.
-    /// T?t c? CRUD ch?m DB ch? ?i qua repository (DAL).
+    /// Triển khai mặc định cho IGenericService sử dụng IGenericRepository.
+    /// Tất cả CRUD chuyển xuống repository (DAL).
     /// </summary>
     public class GenericService<TEntity> : IGenericService<TEntity> where TEntity : class
     {
@@ -18,10 +19,25 @@ namespace NT.BLL.Services
             _repository = repository ?? throw new ArgumentNullException(nameof(repository));
         }
 
+        /// <summary>
+        /// Lấy tất cả entity.
+        /// </summary>
         public Task<IEnumerable<TEntity>> GetAllAsync() => _repository.GetAllAsync();
 
-        public Task<TEntity?> GetByIdAsync(Guid id) => _repository.GetByIdAsync(id);
+        /// <summary>
+        /// Lấy entity theo id.
+        /// </summary>
+        /// <exception cref="ArgumentException">Nếu id == Guid.Empty</exception>
+        public Task<TEntity?> GetByIdAsync(Guid id)
+        {
+            if (id == Guid.Empty) throw new ArgumentException("id không hợp lệ", nameof(id));
+            return _repository.GetByIdAsync(id);
+        }
 
+        /// <summary>
+        /// Thêm entity mới.
+        /// </summary>
+        /// <exception cref="ArgumentNullException">Nếu entity là null</exception>
         public async Task<TEntity> AddAsync(TEntity entity)
         {
             if (entity is null) throw new ArgumentNullException(nameof(entity));
@@ -29,6 +45,10 @@ namespace NT.BLL.Services
             return entity;
         }
 
+        /// <summary>
+        /// Cập nhật entity.
+        /// </summary>
+        /// <exception cref="ArgumentNullException">Nếu entity là null</exception>
         public async Task<TEntity> UpdateAsync(TEntity entity)
         {
             if (entity is null) throw new ArgumentNullException(nameof(entity));
@@ -36,8 +56,14 @@ namespace NT.BLL.Services
             return entity;
         }
 
+        /// <summary>
+        /// Xóa entity theo id.
+        /// Trả về true nếu xóa thành công, false nếu có lỗi (không throw).
+        /// </summary>
         public async Task<bool> DeleteAsync(Guid id)
         {
+            if (id == Guid.Empty) throw new ArgumentException("id không hợp lệ", nameof(id));
+
             try
             {
                 await _repository.DeleteAsync(id);
@@ -45,8 +71,24 @@ namespace NT.BLL.Services
             }
             catch
             {
+                // swallow exception to keep contract of returning bool.
+                // If caller needs exception detail, use repository directly or change contract.
                 return false;
             }
+        }
+
+        /// <summary>
+        /// Lưu tất cả thay đổi pending xuống DB (delegates to repository).
+        /// </summary>
+        public Task SaveChangesAsync() => _repository.SaveChangesAsync();
+
+        /// <summary>
+        /// Tìm kiếm theo biểu thức predicate (delegates to repository).
+        /// </summary>
+        public Task<IEnumerable<TEntity>> FindAsync(Expression<Func<TEntity, bool>> predicate)
+        {
+            if (predicate is null) throw new ArgumentNullException(nameof(predicate));
+            return _repository.FindAsync(predicate);
         }
     }
 }
